+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Add TableView::find_first<T>().
 * Make Table::find_first<T>() public and add support for most column types.
 * Add wrappers for Table::set<T>() to Row.
+* Add support for all column types in Table::get<T>().
 
 -----------